Organize core backend files
