+++ tests/backend/test_pieces.py
